+++ grapher_lib/gui_callbacks_graph_viz.py
@@ -10,6 +10,8 @@ This code is distributed under the MIT License. For more details, see the LICENS
 import polars as pl
 from dash import Output, Input, State, callback, callback_context, no_update
 from grapher_lib import utils as gu
+from grapher_lib import utils_file_upload as fu
+
 
 @callback(
     Output("graphviz-dot", "style"),
@@ -37,12 +39,13 @@ def change_dot_editor_visibility(enable_edit, editor_style):
     Input("checkbox-viz-all-columns", "value"),  # parinktis per Viz grafiko kontekstinį meniu stulpelių rodymui
     Input("checkbox-viz-description", "value"),  # parinktis per Viz grafiko kontekstinį meniu aprašų rodymui
     Input("checkbox-viz-show-checkbox", "value"),  # parinktis per Viz grafiko kontekstinį meniu langelių rodymui
+    Input("memory-viz-imported-checkbox", "data"),
     State("memory-viz-clicked-checkbox", "data"),
     config_prevent_initial_callbacks=True,
 )
 def get_network_viz_chart(
     data_submitted, filtered_elements, engine, layout,
-    show_all_columns, show_descriptions, show_checkbox, viz_selection_dict
+    show_all_columns, show_descriptions, show_checkbox, viz_uploaded_checkboxes, viz_selection_dict
 ):
     """
     Atvaizduoja visas pasirinktas lenteles kaip tinklo mazgus.
@@ -57,12 +60,13 @@ def get_network_viz_chart(
     :param show_all_columns: ar rodyti visus lentelės stulpelius (True); ar tik turinčius ryšių (False)
     :param show_descriptions: ar rodyti lentelių ir stulpelių aprašus pačiame grafike
     :param show_checkbox: ar prie stulpelių pridėti žymimuosius langelius
-    :param viz_selection_dict: Visų sužymėtų langelių simboliai žodyne,
+    :param viz_uploaded_checkboxes: visų sužymėtų langelių simboliai žodyne,
         kur pirmasis lygis yra lentelės, antrasis – stulpeliai, pvz:
         {
             "Skaitytojas": {"ID": "⬜"},
             "Rezervacija": {"ClientID": "🟩", "BookCopyID": "🟥"}}
         }
+    :param viz_selection_dict: ta pati struktūra kaip `viz_uploaded_checkboxes`, tačiau skiriasi jos kilmė
     :return:
     """
     if (engine != "Viz") or (not filtered_elements):
@@ -83,6 +87,8 @@ def get_network_viz_chart(
         df_col = df_nodes_col.filter(pl.col("table").is_in(nodes))
     else:  # Veikti net jei PDSA stulpelius aprašančiame lakšte "table" stulpelio nebūtų
         df_col = pl.DataFrame({"table": {}})  # get_graphviz_dot() sukurs automatiškai pagal ryšius, jei jie yra
+    if viz_uploaded_checkboxes:  # Jei importuoji nauji langelių žymenys iš JSON, naudoti juos
+        viz_selection_dict = viz_uploaded_checkboxes
     df_checkbox = gu.convert_nested_dict2df(viz_selection_dict, ["table", "column", "checkbox"])
     if "checkbox" in df_col:
         df_col = df_col.drop("checkbox")  # išmesti seną stulpelį, nes prijungsim naujas reikšmes iš df_checkbox
@@ -149,13 +155,44 @@ def copy_viz_displayed_nodes_to_clipboard_quoted(filtered_elements, n_clicks):  
 
 
 @callback(
-    Output("memory-viz-clicked-checkbox", "data"),
-    Input("memory-submitted-data", "data"),
-    Input("viz-clicked-checkbox-store", "data"),
-    State("memory-viz-clicked-checkbox", "data"),
+    Output("memory-viz-imported-checkbox", "data"),
+    Input("upload-data-viz-checkbox", "contents"),
+    State("upload-data-viz-checkbox", "filename"),
     prevent_initial_callbacks=True,
 )
-def remember_viz_clicked_checkbox(data_submitted, viz_last_clicked_checkbox, viz_selection_dict):
+def import_checkbox_markings(uploaded_content, uploaded_filenames):
+    """
+    Žymimųjų langelių reikšmių importavimas
+    :param uploaded_content: įkeltų rinkmenų turinys sąrašo pavidalu, kur
+        vienas elementas – vienos rinkmenos base64 turinys
+    :param uploaded_filenames: įkeltos rinkmenos vardas (nors vienas, bet turi būti sąraše)
+    """
+    if uploaded_content:
+        parse_output = fu.parse_file(uploaded_content, uploaded_filenames)
+        if (
+                isinstance(parse_output, dict) and ("file_data" in parse_output) and (
+                "columns" in parse_output["file_data"]) and
+                ("df" in parse_output["file_data"]["columns"]) and parse_output["file_data"]["columns"]["df"]
+        ):
+            df_checkboxes = pl.DataFrame(parse_output["file_data"]["columns"]["df"], infer_schema_length=None)
+            df_checkboxes = fu.select_renamed_or_add_columns(
+                df_checkboxes, old_columns=None, new_columns=["table", "column", "checkbox"]
+            )
+            return gu.convert_df2nested_dict(df_checkboxes, col_names=["table", "column", "checkbox"])
+    return {}
+
+
+@callback(
+    Output("memory-viz-clicked-checkbox", "data"),
+    Input("memory-submitted-data", "data"),
+    Input("viz-clicked-checkbox-store", "data"), # Grafike naudotojo nuspaustas paskutinis langelis
+    Input("memory-viz-imported-checkbox", "data"),  # Langelių žymenys, importuoti iš JSON
+    State("memory-viz-clicked-checkbox", "data"),  # Grafike naudotojo suspaudyti langeliai (visi)
+    prevent_initial_callbacks=True,
+)
+def remember_viz_clicked_checkbox(
+        data_submitted, viz_last_clicked_checkbox, viz_uploaded_checkboxes, viz_selection_dict
+):
     """
     Paskutinio pakeisto žymimojo langelio simbolį įtraukia į visų pakeistųjų žodyną
     :param data_submitted: žodynas su PDSA ("node_data") ir ryšių ("edge_data") duomenimis
@@ -167,6 +204,7 @@ def remember_viz_clicked_checkbox(data_submitted, viz_last_clicked_checkbox, viz
             "symbol": "🟩",
             "parentPosition": {"x": 168.6, "y": 268.6, "width": 275.4, "height": 21.5}
         }
+    :param viz_uploaded_checkboxes: toks pat formatas kaip `viz_selection_dict`, tik atėjęs iš JSON.
     :param viz_selection_dict: Visų sužymėtų langelių simboliai žodyne,
         kur pirmasis lygis yra lentelės, antrasis – stulpeliai, pvz:
         {
@@ -175,8 +213,10 @@ def remember_viz_clicked_checkbox(data_submitted, viz_last_clicked_checkbox, viz
         }
     :return:
     """
-    changed_id = [p["prop_id"] for p in callback_context.triggered][0]  # Sužinoti, kas iškvietė f-ją
-    if (changed_id == "memory-submitted-data.data") and data_submitted:
+    # Sužinoti, kas iškvietė f-ją. Jei po JSON importavimo naudotojas atnaujina puslapį, gali rodyti du:
+    # ['upload-data-viz-checkbox.contents', 'memory-submitted-data.data'] bet šiuo atveju reiktų žiūrėti tik antrąjį
+    changed_ids = [p["prop_id"] for p in callback_context.triggered]
+    if ("memory-submitted-data.data" in changed_ids) and data_submitted:
         # Įkelti nauji duomenys "col_sheet_data"
         if data_submitted["node_data"]["col_sheet_renamed_cols"]["checkbox"]:  # Ar buvo "checkbox" prasmę turintis stulpelis
             return gu.convert_df2nested_dict(
@@ -184,21 +224,25 @@ def remember_viz_clicked_checkbox(data_submitted, viz_last_clicked_checkbox, viz
                 col_names=["table", "column", "checkbox"]
             )
         return {}
-    if not viz_selection_dict:
-        viz_selection_dict = {}
-    if (
-        (not viz_last_clicked_checkbox) or (not isinstance(viz_last_clicked_checkbox, dict)) or
-        ("id" not in viz_last_clicked_checkbox)
-    ):
-        # viz_last_clicked_checkbox netinkamas
-        return no_update
-    id_parts = viz_last_clicked_checkbox["id"].split(":")
-    if len(id_parts) != 2:
-        return no_update
-    table, column = id_parts
-    if table not in viz_selection_dict:
-        viz_selection_dict[table] = {}
-    viz_selection_dict[table][column] = viz_last_clicked_checkbox["symbol"]
+    elif "viz-clicked-checkbox-store.data" in changed_ids:
+        if not viz_selection_dict:
+            viz_selection_dict = {}
+        if (
+            (not viz_last_clicked_checkbox) or (not isinstance(viz_last_clicked_checkbox, dict)) or
+            ("id" not in viz_last_clicked_checkbox)
+        ):
+            # viz_last_clicked_checkbox netinkamas
+            return no_update
+        id_parts = viz_last_clicked_checkbox["id"].split(":")
+        if len(id_parts) != 2:
+            return no_update
+        table, column = id_parts
+        if table not in viz_selection_dict:
+            viz_selection_dict[table] = {}
+        viz_selection_dict[table][column] = viz_last_clicked_checkbox["symbol"]
+    elif ("memory-viz-imported-checkbox.data" in changed_ids) and viz_uploaded_checkboxes:
+        # Naudotojas importavo žymimųjų langelių reikšmes iš JSON per grafiko kontekstinį meniu
+        return viz_uploaded_checkboxes
     return viz_selection_dict